requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
httpx==0.27.0

# LangChain + LangGraph core
langchain==0.2.14
langgraph==0.2.15
langchain-openai==0.1.21

# Storage / DB
asyncpg==0.29.0
supabase==2.5.1
python-dotenv==1.0.1
